Converted to TypeORM and TypeScript implementation
