+++ src/controllers/ArticleController.ts
@@ -1,0 +1,1 @@
+import { Request, Response } from 'express';
