+++ src/controllers/index.html
@@ -1,0 +1,50 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+    <title>Seafarers' Cafe</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <link rel="stylesheet" type="text/css" href="/css/styles.css">
+    <link rel="shortcut icon" href="/favicon.ico">
+    <script src="/js/feed.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/vue"></script>
+    <script>
+        window.onload = loadFeed(function () {
+            var feed = new Vue({
+                el: '.feed',
+                data: {
+                    posts: []
+                },
+                created() {
+                    fetch(new Request('/posts/index')).then(response => response.json())
+                        .then(response => this.posts = response);
+                }
+            });
+        })
+    </script>
+</head>
+
+<body>
+    <div class="titlebar">
+        <h1>Seafarers' Cafe</h1>
+        <nav>
+            <a href="/" class="btn btn-primary">Home</a>
+            <a href="/blog" class="btn btn-primary">Blog</a>
+            <a href="http://podcast.seafarers.cafe" class="btn btn-primary">Podcast</a>
+            <a href="/aboutus" class="btn btn-primary">About Us</a>
+            <a href="/email" class="btn btn-primary">Newsletter</a>
+        </nav>
+    </div>
+    
+
+    <div id="feed"></div>
+
+    <footer>
+        <div>Seafarers' Cafe</div>
+        <div></div><a href="https://github.com/Mark-Powers/marks.kitchen">Web engine built upon an existing kitchen</a></div>
+        <div><a href="/login">Login</a></div>
+    </footer>
+</body>
+
+</html>
