+++ src/entity/media/ArticleType.ts
@@ -1,5 +1,5 @@
 import {Entity, PrimaryGeneratedColumn, Column, OneToMany} from "typeorm";
-import { Article } from "./articles";
+import { Article } from "./Articles";
 
 @Entity()
 export class ArticleType {
