--- src/public/css/bootstrap.css
@@ -1,1 +1,0 @@
-../../node_modules/bootstrap/dist/css/bootstrap.css
