+++ tsconfig.json
@@ -6,6 +6,7 @@
       ],
       "target": "es5",
       "module": "commonjs",
+      "esModuleInterop": true,
       "moduleResolution": "node",
       "outDir": "./build",
       "emitDecoratorMetadata": true,
